updates to compile
